backend/requirements.txt
Flask==2.0.1
Werkzeug==2.2.2
psycopg2-binary
python-dotenv